Move component back to Logout,
Turns out its more of a component than a helper function
