+++ src/components/Logout/index.tsx
@@ -1,4 +1,6 @@
-import { customFetch } from '@opensrp/server-service';
+import { logout } from '@opensrp/server-logout';
+import React from 'react';
+import { useHistory } from 'react-router';
 import {
   BACKEND_ACTIVE,
   DOMAIN_NAME,
@@ -6,24 +8,21 @@ import {
   KEYCLOAK_LOGOUT_URL,
   OPENSRP_LOGOUT_URL,
 } from '../../configs/env';
+import { HOME_URL } from '../../constants';
 import { displayError } from '../../helpers/errors';
-import { getPayloadOptions, OpenSRPService } from '../../services/opensrp';
+import { getPayloadOptions } from '../../services/opensrp';
+import Ripple from '../page/Loading';
 
-/** custom function that logs user from the authorization server */
-export const logoutFromAuthServer = () => {
+/** HOC function that calls function that logs out the user from both opensrp
+ * and keycloak
+ */
+export const customLogout = () => {
+  const history = useHistory();
   const payload = getPayloadOptions(new AbortController().signal, 'GET');
   const redirectUri = BACKEND_ACTIVE ? EXPRESS_OAUTH_LOGOUT_URL : DOMAIN_NAME;
-  const filterParams = {
-    redirect_uri: redirectUri,
-  };
-  const openSRPLogoutUri = OpenSRPService.getURL(OPENSRP_LOGOUT_URL, {});
-  const keycloakLogoutUri = OpenSRPService.getURL(KEYCLOAK_LOGOUT_URL, filterParams);
-  customFetch(openSRPLogoutUri, payload)
-    .then(() => {
-      window.location.href = keycloakLogoutUri;
-    })
-    .catch(error => {
-      displayError(error);
-    });
-  return null;
+  logout(payload, OPENSRP_LOGOUT_URL, KEYCLOAK_LOGOUT_URL, redirectUri).catch(error => {
+    displayError(error);
+    history.push(HOME_URL);
+  });
+  return <Ripple />;
 };
